pyproject.toml
[project]
name = "utec-planificador-agent"
version = "0.1.0"
description = "AI agent microservice for UTEC Planificador Docente"
requires-python = ">=3.9"

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.0.20",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.2",
    "loguru>=0.7.0",
    "pydantic>=2.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

